requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.8.0
exceptiongroup==1.2.0
fastapi==0.104.1
frozenlist==1.4.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.6
jq==1.6.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.341
langchain-core==0.0.6
langchain-experimental==0.0.42
langsmith==0.0.67
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.2.2
openai==1.3.5
packaging==23.2
pgvector==0.2.4
priority==2.0.0
proto-plus==1.22.3
protobuf==4.25.1
psycopg==3.1.13
psycopg-binary==3.1.13
psycopg-pool==3.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
taskgroup==0.0.0a4
tenacity==8.2.3
tiktoken==0.5.1
tomli==2.0.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.3
